rangekey: move iteration stack to rangekeystack
