+++ internal/rangekeystack/testdata/defragmenting_iter
